1st_project/requirements.txt
numpy==1.24.2
pdfreader==0.1.12
PyPDF2==3.0.1
tabula-py==2.7.0
